requirements/requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
Pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
